% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.median.track.R
\name{geo.median.track}
\alias{geo.median.track}
\title{Function to recalculate track with possibilities in non-active hemisphere removed}
\usage{
geo.median.track(pr, cross.north = NULL, cross.south = NULL)
}
\arguments{
\item{cross.north}{The expected date that a subject would cross the equator to the northern hemisphere. Character string in the form of "2022-04-15". NULL if a northward crossing isn't expected.}

\item{all.tracks}{First element of output from GLS.prob.algorithm function. All possible tracks.}

\item{cross.southThe}{The expected date that a subject would cross the equator to the southern hemisphere. Character string in the form of "2022-04-15". NULL if a southward crossing isn't expected.}
}
\description{
Function to recalculate track with possibilities in non-active hemisphere removed
}
