% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.land.mask.R
\name{modify.land.mask}
\alias{modify.land.mask}
\title{Define land mask used in prob.GLS.algorithm}
\usage{
modify.land.mask(
  med.sea = F,
  black.sea = F,
  baltic.sea = F,
  caspian.sea = F,
  arctic.ocean = F,
  north.atlantic.ocean = F,
  south.atlantic.ocean = F,
  north.pacific.ocean = F,
  south.pacific.ocean = F,
  southern.ocean = F,
  custom.land.mask = NULL
)
}
\arguments{
\item{med.sea}{if T classify mediterranean sea as land}

\item{black.sea}{if T classify black sea as land}

\item{baltic.sea}{if T classify baltic sea as land}

\item{caspian.sea}{if T classify caspian sea as land}

\item{arctic.ocean}{if T classify arctic ocean as land}

\item{north.atlantic.ocean}{if T classify north atlantic ocean as land}

\item{south.atlantic.ocean}{if T classify south atlantic ocean as land}

\item{north.pacific.ocean}{if T classify north pacific ocean as land}

\item{south.pacific.ocean}{if T classify south pacific ocean as land}

\item{southern.ocean}{if T classify southern ocean as land}

\item{custom.land.mask}{List containing vectors of length four that define custom areas to include in land mask. The format is minimum.longitude, maximum.longitude, minimum latitude, maximum latitude.}

\item{NOAA.OI.location}{The location of the NOAA OI SST V2 land mask file relative to the working directory.}
}
\description{
Extends land mask found in NOAA OI SST V2 file. Land masks mark the boundaries where position estimates can not be made.
}
