% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLS.prob.algorithm.R
\name{plot.glsTracks}
\alias{plot.glsTracks}
\title{Generate series of plots displaying iteration location estimates at each timestep overlaid with most probable track.}
\usage{
plot.glsTracks(obj, zoom = TRUE, ...)
}
\arguments{
\item{obj}{object of class "glsTracks"}

\item{zoom}{If T plot limits are min/max of all iteration estimates. If F plot limits are bounding box.}

\item{...}{additional arguments to be passed to method.}
}
\description{
Generate series of plots displaying iteration location estimates at each timestep overlaid with most probable track.
}
