% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLS.prob.algorithm.R
\name{print.summary.tracks}
\alias{print.summary.tracks}
\title{Print method for summary(glsTracks)}
\usage{
print.summary.tracks(obj, ...)
}
\arguments{
\item{obj}{object of class "summary.glsTracks"}

\item{...}{additional arguments to be passed to methods}
}
\description{
Print method for summary(glsTracks)
}
