% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.act.R
\name{read.act}
\alias{read.act}
\title{Read in data from .act files}
\usage{
read.act(actfile, sampling.interval = 3, summarise.interval = 10)
}
\arguments{
\item{actfile}{The name of an activity file downloaded from a geolocator device.}

\item{sampling.interval}{The interval at which the device samples state changes in seconds (e.g. MK15 & MK3006 sample every 3 sec).}

\item{summarise.interval}{The number of minutes to sum all the wet state changes. Defaults to 10 minutes.}
}
\description{
Transforms .act files into a data frame with datetime and wetdry values. Wetdry value represents the total number of times that the sensor is recorded as submerged in salt water within a time interval.
}
