% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.sensor.R
\name{read.sensor}
\alias{read.sensor}
\title{Read in data from .tem or .sst files from MK7, MK15 or C65-SUPER GLS tags.}
\usage{
read.sensor(temfile, logger.type = "MK15", temp.range = c(-2, 19))
}
\arguments{
\item{temfile}{The name of a temperature log file downloaded from a geolocator device}

\item{logger.type}{The model of light-level geolocator. BAS MK7, MK15 and Migrate Technology C65-SUPER are supported.}

\item{temp.range}{Vector containing to lowest and highest temperature values. Migrate technology loggers provide min and max temperatures for each eight hour interval.}
}
\description{
Transforms .tem or .sst files into a data frame consistent with the requirements of GLS.prob.algorithm or prob_algorithm
}
