% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.trn.R
\name{read.trn}
\alias{read.trn}
\title{Read in data from BAS/Biotrack devices}
\usage{
read.trn(trnfile)
}
\arguments{
\item{trnfile}{The name of a file downloaded from a BAS/Biotrack device}
}
\description{
Transforms biotrack or BAS .trn files into a data frame with tFirst, tSecond and type.
}
\examples{
trn           <- twilightCalc(BBA_lux$dtime, BBA_lux$lig, ask = FALSE, LightThreshold = 2)
summary(trn)

}
