% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLS.prob.algorithm.R
\name{summary_glsTracks}
\alias{summary_glsTracks}
\title{Summary method for glsTracks object}
\usage{
summary_glsTracks(obj, ...)
}
\arguments{
\item{obj}{object of class "glsTracks"}

\item{...}{additional arguments to be passed to methods}
}
\description{
Summary method for glsTracks object
}
