% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.land.mask.R
\name{summary_landmask}
\alias{summary_landmask}
\title{Summary method for landmask object}
\usage{
summary_landmask(obj, ...)
}
\arguments{
\item{obj}{object of class "landmask"}

\item{...}{additional arguments to be passed to methods}
}
\description{
Summary method for landmask object
}
