% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.median.track.R
\name{summary_modifiedTrack}
\alias{summary_modifiedTrack}
\title{Summary method for modifiedTrack object}
\usage{
summary_modifiedTrack(obj, ...)
}
\arguments{
\item{obj}{object of class "modifiedTrack"}

\item{...}{additional arguments to be passed to methods}
}
\description{
Summary method for modifiedTrack object
}
