% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.trn.R
\name{read.trn}
\alias{read.trn}
\title{Read in sunrise/sunset data from BAS/Biotrack devices}
\usage{
read.trn(trnfile)
}
\arguments{
\item{trnfile}{The name of a file downloaded from a BAS/Biotrack device}
}
\description{
Transforms biotrack or BAS .trn files into a data frame with tFirst, tSecond and type. tFirst is the first twilight event of that set, type defines the associated twilight event. (1 for Sunrise, 2 for Sunset) and tSecond is the opposing twilight event. There are not necessarily 2 twilight event recordings per day if the threshold as defined in `Ekstrom, Philip A et al. 2004. “An Advance in Geolocation by Light."` is not met.
}
